requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
html5lib==1.1
PyYAML==6.0.2
python-jobspy
pandas
pytest
